+++ src/types/branchTreeItem.ts
@@ -1,0 +1,6 @@
+export interface IBranchTreeItem {
+    id: number;
+    title: string;
+    children?: IBranchTreeItem[];
+    customIcon?: string;
+  }
