+++ src/types/directory.ts
@@ -1,0 +1,5 @@
+export interface IDirectory {
+    name: string;
+    children: IDirectory[];
+    branches: string[];
+  }
